+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/BloodStalkEntity.java
@@ -2,16 +2,11 @@ package com.kelco.kamenridercraft.entities.bosses;
 
 import com.kelco.kamenridercraft.entities.footSoldiers.BaseHenchmenEntity;
 import com.kelco.kamenridercraft.item.Build_Rider_Items;
-import com.kelco.kamenridercraft.item.Kabuto_Rider_Items;
-import com.kelco.kamenridercraft.item.W_Rider_Items;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.monster.Monster;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 
@@ -31,7 +26,7 @@ public class BloodStalkEntity extends BaseHenchmenEntity {
 	public static AttributeSupplier.Builder setAttributes() {
 		return Monster.createMonsterAttributes()
         		.add(Attributes.FOLLOW_RANGE, 128.0D)
-        		.add(Attributes.MOVEMENT_SPEED,(double)0.30F)
+        		.add(Attributes.MOVEMENT_SPEED, 0.30F)
         		.add(Attributes.ATTACK_DAMAGE, 5.0D)
         		.add(Attributes.MAX_HEALTH, 100.0D);
      }
