+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/CaliburEntity.java
@@ -53,7 +53,7 @@ public class CaliburEntity extends BaseHenchmenEntity {
 
         return Monster.createMonsterAttributes()
         		.add(Attributes.FOLLOW_RANGE, 135.0D)
-        		.add(Attributes.MOVEMENT_SPEED,(double)0.15F)
+        		.add(Attributes.MOVEMENT_SPEED, 0.15F)
         		.add(Attributes.ATTACK_DAMAGE, 5.0D)
         		.add(Attributes.ARMOR, 3.0D)
         		.add(Attributes.MAX_HEALTH, 50.0D);
