+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/ElOfTheWaterEntity.java
@@ -24,7 +24,7 @@ public class ElOfTheWaterEntity extends BaseHenchmenEntity {
     public static AttributeSupplier.Builder setAttributes() {
         return Monster.createMonsterAttributes()
         		.add(Attributes.FOLLOW_RANGE, 135.0D)
-        		.add(Attributes.MOVEMENT_SPEED,(double)0.2F)
+        		.add(Attributes.MOVEMENT_SPEED, 0.2F)
         		.add(Attributes.ATTACK_DAMAGE, 10.0D)
         		.add(Attributes.ARMOR, 3.0D)
         		.add(Attributes.MAX_HEALTH, 60.0D);
