+++ src/main/java/com/kelco/kamenridercraft/entities/bosses/OwlLostSmashEntity.java
@@ -1,26 +1,11 @@
 package com.kelco.kamenridercraft.entities.bosses;
 
-import com.kelco.kamenridercraft.entities.MobsCore;
 import com.kelco.kamenridercraft.entities.footSoldiers.BaseHenchmenEntity;
-import net.minecraft.network.chat.Component;
-import net.minecraft.util.RandomSource;
-import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.MobSpawnType;
-import net.minecraft.world.entity.SpawnGroupData;
 import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.monster.Monster;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.ServerLevelAccessor;
-
-import javax.annotation.Nullable;
-import java.time.LocalDate;
-import java.time.temporal.ChronoField;
 
 public class OwlLostSmashEntity extends BaseHenchmenEntity {
 
@@ -33,7 +18,7 @@ public class OwlLostSmashEntity extends BaseHenchmenEntity {
 
             return Monster.createMonsterAttributes()
                     .add(Attributes.FOLLOW_RANGE, 35.0D)
-                    .add(Attributes.MOVEMENT_SPEED,(double)0.3F)
+                    .add(Attributes.MOVEMENT_SPEED, 0.3F)
                     .add(Attributes.ATTACK_DAMAGE, 10.0D)
                     .add(Attributes.ARMOR, 4.0D)
                     .add(Attributes.MAX_HEALTH, 80.0D);
